requirements-docs.txt
mkdocs>=1.6.0
mkdocs-material>=9.5.0
PyYAML>=6.0.0
